+++ nebuchadnezzar/nebu/cli/main.py
@@ -1,13 +1,20 @@
 # -*- coding: utf-8 -*-
 """Commandline utility for publishing content"""
+import os
 import shutil
+import sys
 import tempfile
 import zipfile
 from pathlib import Path
 
 import click
 import requests
-from litezip import convert_completezip, parse_litezip, validate_litezip
+from litezip import (
+    Collection,
+    convert_completezip,
+    parse_litezip,
+    validate_litezip,
+)
 
 from ..logger import configure_logging, logger
 
@@ -139,12 +146,12 @@ def get(col_id, col_version, output_dir):
     shutil.rmtree(str(tmp_dir))
 
 
-@cli.command()
-@click.argument('content_dir', default='.',
-                type=click.Path(exists=True, file_okay=False))
-def validate(content_dir):
-    content_dir = Path(content_dir).resolve()
-    struct = parse_litezip(content_dir)
+def is_valid(struct):
+    """Checks validity of a litezip's contents with commandline ouput
+    containing any validation errors.
+    Returns True when the content is value and False if not.
+
+    """
 
     has_errors = False
     cwd = Path('.').resolve()
@@ -152,8 +159,103 @@ def validate(content_dir):
         has_errors = True
         filepath = filepath.relative_to(cwd)
         logger.error('{}:{}'.format(filepath, error_msg))
+    return not has_errors
 
-    if has_errors:
-        logger.info("We've got problems... :(")
-    else:
+
+@cli.command()
+@click.argument('content_dir', default='.',
+                type=click.Path(exists=True, file_okay=False))
+def validate(content_dir):
+    content_dir = Path(content_dir).resolve()
+    struct = parse_litezip(content_dir)
+
+    if is_valid(struct):
         logger.info("All good! :)")
+    else:
+        logger.info("We've got problems... :(")
+
+
+def _publish(struct, message):
+    """Publish the struct to a repository"""
+    collection_id = struct[0].id
+    # Base encapsulating directory within the zipfile
+    base_file_path = Path(collection_id)
+
+    # Zip it up!
+    # TODO Move this block of logic to litezip. Maybe?
+    _, zip_file = tempfile.mkstemp()
+    zip_file = Path(zip_file)
+    with zipfile.ZipFile(str(zip_file), 'w') as zb:
+        for model in struct:
+            # Write the content file into the zip.
+            if isinstance(model, Collection):
+                file = model.file
+                rel_file_path = base_file_path / model.file.name
+            else:  # Module
+                file = model.file
+                rel_file_path = base_file_path / model.id / model.file.name
+            zb.write(str(file), str(rel_file_path))
+            # TODO Include resource files
+
+    # Send it!
+    # FIXME We need to be able to build urls to multiple services.
+    #       For now we'll use an environment variable
+    host = os.environ.get('XXX_HOST', 'cnx.org')
+    url = 'https://{}/api/v3/publish'.format(host)
+    # FIXME We don't have nor want explicit setting of the publisher.
+    #       The publisher will come through as part of the authentication
+    #       information, which will be in a later implementation.
+    #       For now, pull it out of a environment variable.
+    data = {
+        'publisher': os.environ.get('XXX_PUBLISHER', 'OpenStaxCollege'),
+        'message': message,
+    }
+    files = {
+        'file': ('contents.zip', zip_file.open('rb'),),
+    }
+    resp = requests.post(url, data=data, files=files)
+
+    # Clean up!
+    zip_file.unlink()
+
+    # Process any response messages
+    if resp.status_code == 200:
+        # TODO Nicely format this stuff. Wait for the Bravado client
+        #      implementation to work with models to make this work easier.
+        logger.debug('Temporary raw output...')
+        from pprint import pformat
+        logger.info('Publishing response: \n{}'
+                    .format(pformat(resp.json())))
+    elif resp.status_code == 400:
+        # This way be Errors
+        # TODO Nicely format this stuff. Wait for the Bravado client
+        #      implementation to work with models to make this work easier.
+        logger.debug('Temporary raw output...')
+        logger.error('ERROR: \n{}'.format(resp.content))
+        return False
+    else:  # pragma: no cover
+        logger.error("unknown response:\n  status: {}\n  contents: {}"
+                     .format(resp.status_code, resp.text))
+        raise RuntimeError('unknown response, see output above')
+
+    return True
+
+
+@cli.command()
+@click.argument('content_dir',
+                type=click.Path(exists=True, file_okay=False))
+@click.argument('publication_message', type=str)
+def publish(content_dir, publication_message):
+    content_dir = Path(content_dir).resolve()
+    struct = parse_litezip(content_dir)
+
+    if is_valid(struct):
+        has_published = _publish(struct, publication_message)
+        if has_published:
+            logger.info("Great work!!! =D")
+        else:
+            logger.info("Stop the Press!!! =()")
+            sys.exit(1)
+    else:
+        logger.info("We've got problems... :(")
+        sys.exit(1)