+++ nebuchadnezzar/docs/changes.rst
@@ -5,4 +5,4 @@ Change Log
 ?.?.?
 -----
 
-- ?
+- Adds a publish command that communicates with a Press service.